script_push

+++ shell_script/git_push_master.sh
@@ -3,7 +3,10 @@ cd "/mnt/d/Google ドライブ/Pipeline"
 source "/mnt/d/Google ドライブ/00Colab Notebooks/00Commands/03SCRIPTkey/github.txt"
 git checkout master
 #gitignoreにtxtついてないか注意
-git rm -r --cached .
+if [ $1 ]; then
+  git rm -r --cached .
+fi
+
 git add -A
 git commit -m "script_push"
